Split object header into two components: ObjectItemsHeader and ObjectDetailsHeader
