+++ frontend/app/src/entities/proposed-changes/ui/proposed-changes-manager-toolbar.tsx
@@ -6,7 +6,7 @@ import { ButtonWithTooltip } from "@/shared/components/buttons/button-primitive"
 
 import { ActiveObjectFilterTags } from "@/entities/nodes/object/ui/filters/active-object-filter-tags";
 import { FilterSearchInput } from "@/entities/nodes/object/ui/filters/filter-search-input";
-import { ObjectItemsHeader } from "@/entities/nodes/object/ui/object-header";
+import { ObjectItemsHeader } from "@/entities/nodes/object/ui/object-items-header";
 import type { Permission } from "@/entities/permission/types";
 import type { ModelSchema } from "@/entities/schema/types";
